+++ src/test/java/io/apimap/cli/PublishTest.java
@@ -27,12 +27,5 @@ import java.io.IOException;
 public class PublishTest {
     @Test
     public void publishAPI_didSucceed() throws IOException {
-        PublishCommand cmd = new PublishCommand("examples/metadata.apicatalog",
-            "examples/taxonomy.apicatalog",
-                "http://localhost:8081",
-                "git://",
-                null);
-
-        cmd.run();
     }
 }