+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 Changelog
 ===
 
+### 26.08.2022
+- Set version number from argument
+
+### 09.08.2022
+- Added README.md and CHANGELOG.md support
+- Updated dependencies 
+
 ### 21.04.2022
-- asd
+- Added action confirm dialogs on dangerous actions
+- Fixed token file bug
+- Updated documentation
+- Added static endpoint configuration if only one apimap instance is to be used
+- Simplified token list
 
 ### 05.11.2021 
 - First public release